NO-TICKET: Move ByteString Prettifier to shared.
Add cats.Show instance for ByteString.
